src/spark/requirements/requirements.txt
ipython
pandas
pyarrow
numpy
pyspark
kafka-python